moved node types to mission
