+++ src/components/waypointList/MissionList.tsx
@@ -2,12 +2,12 @@ import { useWaypoints } from "@/util/context/WaypointContext";
 import CreateCollection from "./createCollection";
 import CollectionItem from "./collectionItem";
 import ListItem from "./ListItem";
-import { CollectionType } from "@/types/waypoints";
 import { TfiTarget } from "react-icons/tfi";
 import { commandName } from "@/util/translationTable";
 import { FaArrowRight, FaTrashAlt } from "react-icons/fa";
 import { useState } from "react";
 import { commands } from "@/lib/commands/commands";
+import { CollectionType } from "@/lib/mission/mission";
 
 export default function MissionList({ onHide }: { onHide: () => void }) {
   const { setActiveMission, waypoints, setSelectedWPs, selectedWPs, setWaypoints, activeMission, setTool } = useWaypoints()
