+++ src/components/waypointList/collectionItem.tsx
@@ -1,9 +1,9 @@
-import { ColNode } from "@/types/waypoints";
 import { useWaypoints } from "@/util/context/WaypointContext";
 import { useState } from "react";
 import { FaPlus, FaTrashAlt } from "react-icons/fa";
 import { TbTopologyRing } from "react-icons/tb";
 import ListItem from "./ListItem";
+import { ColNode } from "@/lib/mission/mission";
 
 export default function CollectionItem({ node, selected, onMouseDown, onDelete }: { node: ColNode, selected: boolean, onMouseDown: (e: React.MouseEvent<HTMLDivElement>) => void, onDelete: () => void }) {
   const [expand, setExpand] = useState(false)
