+++ src/lib/mission/tests/WaypointCollection.test.ts
@@ -1,6 +1,5 @@
 import { makeCommand } from "@/lib/commands/default";
-import { MissingMission, RecursiveMission, Mission } from "@/lib/mission/mission";
-import { CollectionType } from "@/types/waypoints";
+import { MissingMission, RecursiveMission, Mission, CollectionType } from "@/lib/mission/mission";
 import { expect, test } from "bun:test";
 
 test("WaypointCollection default config", () => {
